city-dashboard/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
awscrt==0.32.1
awsiotsdk==1.29.0
boto3==1.42.91
botocore==1.42.91
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
databricks-sql-connector==4.2.5
et_xmlfile==2.0.0
fastapi==0.136.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
jmespath==1.1.0
lz4==4.4.5
numpy==2.4.4
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.3
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybreaker==1.4.1
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
starlette==1.0.0
thrift==0.20.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
